This is reading from file & sorting the data
This sorts the data from certain ranges & tallies them up within those groups.
Write a program that reads a file consisting of students’ test scores in the range 0–200. It should then determine the number of students having scores in each of the following ranges:
0–24, 25–49, 50–74, 75–99, 100–124, 125–149, 150–174, and 175–200.
Output the score ranges and the number of students. (Run your program with the following input data:

76, 89, 150, 135, 200, 76, 12, 100, 150, 28, 178, 189, 167, 
200, 175, 150, 87, 99, 129, 149, 176, 
200, 87, 35, 157, 189